reactor view related actions and move them to ViewActions
https://github.com/metasfresh/metasfresh-webui-frontend/issues/1407
